# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: SSCMS API
  description: 'Smart Student Club Management System -- Backend Sprint 1 (Milestone
    3) covered Epics 1-3: Role-Based Portal Access & Onboarding, Venue Booking & Asset
    Management, and Event Lifecycle & Approval Workflows. Backend Sprint 2 (Milestone
    4) adds Epics 4-6: Task Management & Volunteer Workflows, QR Attendance & Live
    Execution, and Certification & Budget Analytics.'
  version: 1.1.0
paths:
...
  /clubs:
    get:
      tags:
      - Clubs & Membership
      summary: List Clubs
      operationId: list_clubs_clubs_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ClubOut'
                type: array
                title: Response List Clubs Clubs Get
    post:
      tags:
      - Clubs & Membership
      summary: Create Club
      operationId: create_club_clubs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClubCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubCreateResponse'
        '409':
          description: Club name or club head email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /clubs/{club_id}:
    get:
      tags:
      - Clubs & Membership
      summary: Get Club
      operationId: get_club_clubs__club_id__get
      parameters:
      - name: club_id
        in: path
        required: true
        schema:
          type: integer
          title: Club Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubOut'
        '404':
          description: Club not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Clubs & Membership
      summary: Update Club
      operationId: update_club_clubs__club_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: club_id
        in: path
        required: true
        schema:
          type: integer
          title: Club Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClubUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubOut'
        '403':
          description: Caller cannot manage this club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '404':
          description: Club not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Clubs & Membership
      summary: Delete Club
      operationId: delete_club_clubs__club_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: club_id
        in: path
        required: true
        schema:
          type: integer
          title: Club Id
      responses:
        '204':
          description: Successful Response
        '403':
          description: Caller cannot manage this club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '404':
          description: Club not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /events:
    get:
      tags:
      - Event Proposals & Events
      summary: List Events
      operationId: list_events_events_get
      parameters:
      - name: club_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Club Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventOut'
                title: Response List Events Events Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /events/{event_id}:
    get:
      tags:
      - Event Proposals & Events
      summary: Get Event
      operationId: get_event_events__event_id__get
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventOut'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /reports:
    post:
      tags:
      - Report Generation
      summary: Generate Report
      description: 'Generate and persist an analytics report.  The Faculty Coordinator

        can request a specific report type with an optional date range filter.'
      operationId: generate_report_reports_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOut'
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Unprocessable Entity
    get:
      tags:
      - Report Generation
      summary: List Reports
      description: List all previously generated reports, optionally filtered by type.
      operationId: list_reports_reports_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: report_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Report Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportOut'
                title: Response List Reports Reports Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/{report_id}:
    get:
      tags:
      - Report Generation
      summary: Get Report
      description: Retrieve a single report by ID.
      operationId: get_report_reports__report_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: integer
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/{report_id}/download:
    get:
      tags:
      - Report Generation
      summary: Download Report
      description: Download a report as a CSV file.
      operationId: download_report_reports__report_id__download_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: integer
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ClubCreate:
      properties:
        name:
          type: string
          title: Name
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        history:
          anyOf:
          - type: string
          - type: 'null'
          title: History
        achievements:
          items:
            type: string
          type: array
          title: Achievements
          default: []
        club_head_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Club Head Email
          description: Defaults to <clubname-slug>@iitm.in if omitted
        club_head_password:
          anyOf:
          - type: string
            minLength: 6
          - type: 'null'
          title: Club Head Password
          description: Defaults to <clubname-slug>123 if omitted
      type: object
      required:
      - name
      title: ClubCreate
    ClubCreateResponse:
      properties:
        club:
          $ref: '#/components/schemas/ClubOut'
        club_head:
          $ref: '#/components/schemas/ClubHeadCredentialsOut'
      type: object
      required:
      - club
      - club_head
      title: ClubCreateResponse
...
    ClubHeadCredentialsOut:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        note:
          type: string
          title: Note
          default: Share these with the incoming Club Head -- this is the only time
            the password is returned in plaintext.
      type: object
      required:
      - email
      - password
      title: ClubHeadCredentialsOut
...
    ClubOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        history:
          anyOf:
          - type: string
          - type: 'null'
          title: History
        achievements:
          items:
            type: string
          type: array
          title: Achievements
          default: []
        status:
          type: string
          title: Status
        president_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: President Id
        faculty_coordinator_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Faculty Coordinator Id
        budget_allotted:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Budget Allotted
          default: '0'
        budget_spent:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Budget Spent
          default: '0'
      type: object
      required:
      - id
      - name
      - status
      title: ClubOut
...
    ClubUpdate:
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        history:
          anyOf:
          - type: string
          - type: 'null'
          title: History
      type: object
      title: ClubUpdate
...
    ErrorOut:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorOut
...
    EventOut:
      properties:
        id:
          type: integer
          title: Id
        proposal_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Proposal Id
        club_id:
          type: integer
          title: Club Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        venue_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Venue Id
        event_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Event Date
        registration_deadline:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Registration Deadline
        published:
          type: boolean
          title: Published
        finalized:
          type: boolean
          title: Finalized
      type: object
      required:
      - id
      - club_id
      - name
      - published
      - finalized
      title: EventOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReportCreate:
      properties:
        report_type:
          type: string
          title: Report Type
          description: 'One of: club_summary, event_summary, volunteer_summary, finance_summary,
            comprehensive'
        date_from:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date From
        date_to:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date To
      type: object
      required:
      - report_type
      title: ReportCreate
    ReportOut:
      properties:
        id:
          type: integer
          title: Id
        report_type:
          type: string
          title: Report Type
        title:
          type: string
          title: Title
        date_from:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date From
        date_to:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date To
        data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Data
        generated_by:
          type: integer
          title: Generated By
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - id
      - report_type
      - title
      - generated_by
      title: ReportOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login